My commits

+++ 0x14-bit_manipulation/5-flip_bits.c
@@ -11,8 +11,8 @@
 unsigned int flip_bits(unsigned long int n, unsigned long int m)
 {
 	int i, len = 0;
-	unsigned int present;
-	unsigned int exclusive = n ^ m;
+	unsigned long int present;
+	unsigned long int exclusive = n ^ m;
 
 	for (i = 63; i >= 0; i--)
 	{
